hoist selected for plots count
