+++ webview/src/experiments/components/table/TableHead.tsx
@@ -16,6 +16,7 @@ import {
   OnDragOver,
   OnDragStart
 } from '../../../shared/components/dragDrop/DragDropWorkbench'
+import { getSelectedForPlotsCount } from '../../util/rows'
 
 interface TableHeadProps {
   instance: TableInstance<Experiment>
@@ -96,13 +97,15 @@ export const TableHead = ({
     threshold: 1
   })
 
+  const selectedForPlotsCount = getSelectedForPlotsCount(rows)
+
   return (
     <div
       className={cx(styles.thead, needsShadow && styles.headWithShadow)}
       ref={ref}
     >
       <Indicators
-        rows={rows}
+        selectedForPlotsCount={selectedForPlotsCount}
         sorts={sorts}
         filters={filters}
         filteredCounts={filteredCounts}
